requirements.txt
langchain
openai
PyPDF2
faiss-cpu
tiktoken
torch
transformers
sentence-transformers
python-dotenv
streamlit
streamlit_chat
pdfminer
pdfminer.six